backend/app/requirements.txt
Flask==2.2.5
flask_sqlalchemy==3.0.2
flask_cors==3.0.10
networkx==3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
flask_cors